# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Balance API",
        "description": "API для управления балансом пользователей",
        "version": "1.0.0"
    },
    "paths": {
        "/api/balance/{email}": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "Получить баланс пользователя",
                "description": "Возвращает текущий баланс пользователя по email",
                "operationId": "getUserBalance",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "ivan@example.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1000.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Balance",
            "description": "Операции с балансом"
        }
    ]
}